+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Delete tasks 
 - Show loading indicator when tasks are being loading
 - Retrieve tasks from firebase
 - Save tasks on firebase realtime database